Implement light logging

+++ server/UserData/Log/dungeon_log.cs
@@ -1,10 +1,10 @@
 namespace Nanina.UserData.Log;
 
-public class DungeonLog(short[] waifuLVL, short floor, long clearTime, float[] claimDmg)
+public class DungeonLog(short[] waifuLVL, short floor, long clearTime, double[] claimDmg)
 {
     public short floor = floor;
     public short[] waifuLVL { get; set; } = waifuLVL;
     public long timestamp { get; set; } = Utils.GetTimestamp();
     public long clearTime { get; set; } = clearTime;
-    public float[] claimDmg { get; set; } = claimDmg;
+    public double[] claimDmg { get; set; } = claimDmg;
 }
